+++ oci/spec_test.go
@@ -1,5 +1,3 @@
-// +build !windows
-
 /*
    Copyright The containerd Authors.
 
@@ -20,6 +18,7 @@ package oci
 
 import (
 	"context"
+	"runtime"
 	"testing"
 
 	"github.com/containerd/containerd/containers"
@@ -39,26 +38,32 @@ func TestGenerateSpec(t *testing.T) {
 		t.Fatal("GenerateSpec() returns a nil spec")
 	}
 
-	// check for matching caps
-	defaults := defaultCaps()
-	for _, cl := range [][]string{
-		s.Process.Capabilities.Bounding,
-		s.Process.Capabilities.Permitted,
-		s.Process.Capabilities.Inheritable,
-		s.Process.Capabilities.Effective,
-	} {
-		for i := 0; i < len(defaults); i++ {
-			if cl[i] != defaults[i] {
-				t.Errorf("cap at %d does not match set %q != %q", i, defaults[i], cl[i])
+	if runtime.GOOS != "windows" {
+		// check for matching caps
+		defaults := defaultUnixCaps()
+		for _, cl := range [][]string{
+			s.Process.Capabilities.Bounding,
+			s.Process.Capabilities.Permitted,
+			s.Process.Capabilities.Inheritable,
+			s.Process.Capabilities.Effective,
+		} {
+			for i := 0; i < len(defaults); i++ {
+				if cl[i] != defaults[i] {
+					t.Errorf("cap at %d does not match set %q != %q", i, defaults[i], cl[i])
+				}
 			}
 		}
-	}
 
-	// check default namespaces
-	defaultNS := defaultNamespaces()
-	for i, ns := range s.Linux.Namespaces {
-		if defaultNS[i] != ns {
-			t.Errorf("ns at %d does not match set %q != %q", i, defaultNS[i], ns)
+		// check default namespaces
+		defaultNS := defaultUnixNamespaces()
+		for i, ns := range s.Linux.Namespaces {
+			if defaultNS[i] != ns {
+				t.Errorf("ns at %d does not match set %q != %q", i, defaultNS[i], ns)
+			}
+		}
+	} else {
+		if s.Windows == nil {
+			t.Fatal("Windows section of spec not filled on on Windows platform")
 		}
 	}
 
@@ -79,9 +84,15 @@ func TestSpecWithTTY(t *testing.T) {
 	if !s.Process.Terminal {
 		t.Error("terminal net set WithTTY()")
 	}
-	v := s.Process.Env[len(s.Process.Env)-1]
-	if v != "TERM=xterm" {
-		t.Errorf("xterm not set in env for TTY")
+	if runtime.GOOS != "windows" {
+		v := s.Process.Env[len(s.Process.Env)-1]
+		if v != "TERM=xterm" {
+			t.Errorf("xterm not set in env for TTY")
+		}
+	} else {
+		if len(s.Process.Env) != 0 {
+			t.Fatal("Windows process args should be empty by default")
+		}
 	}
 }
 
@@ -91,12 +102,18 @@ func TestWithLinuxNamespace(t *testing.T) {
 	ctx := namespaces.WithNamespace(context.Background(), "testing")
 	replacedNS := specs.LinuxNamespace{Type: specs.NetworkNamespace, Path: "/var/run/netns/test"}
 
-	s, err := GenerateSpec(ctx, nil, &containers.Container{ID: t.Name()}, WithLinuxNamespace(replacedNS))
+	var s *specs.Spec
+	var err error
+	if runtime.GOOS != "windows" {
+		s, err = GenerateSpec(ctx, nil, &containers.Container{ID: t.Name()}, WithLinuxNamespace(replacedNS))
+	} else {
+		s, err = GenerateSpecWithPlatform(ctx, nil, "linux/amd64", &containers.Container{ID: t.Name()}, WithLinuxNamespace(replacedNS))
+	}
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	defaultNS := defaultNamespaces()
+	defaultNS := defaultUnixNamespaces()
 	found := false
 	for i, ns := range s.Linux.Namespaces {
 		if ns == replacedNS && !found {
@@ -114,9 +131,16 @@ func TestWithCapabilities(t *testing.T) {
 
 	ctx := namespaces.WithNamespace(context.Background(), "testing")
 
-	s, err := GenerateSpec(ctx, nil, &containers.Container{ID: t.Name()},
+	opts := []SpecOpts{
 		WithCapabilities([]string{"CAP_SYS_ADMIN"}),
-	)
+	}
+	var s *specs.Spec
+	var err error
+	if runtime.GOOS != "windows" {
+		s, err = GenerateSpec(ctx, nil, &containers.Container{ID: t.Name()}, opts...)
+	} else {
+		s, err = GenerateSpecWithPlatform(ctx, nil, "linux/amd64", &containers.Container{ID: t.Name()}, opts...)
+	}
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -166,13 +190,20 @@ func TestWithPrivileged(t *testing.T) {
 
 	ctx := namespaces.WithNamespace(context.Background(), "testing")
 
-	s, err := GenerateSpec(ctx, nil, &containers.Container{ID: t.Name()},
+	opts := []SpecOpts{
 		WithCapabilities(nil),
 		WithMounts([]specs.Mount{
 			{Type: "cgroup", Destination: "/sys/fs/cgroup", Options: []string{"ro"}},
 		}),
 		WithPrivileged,
-	)
+	}
+	var s *specs.Spec
+	var err error
+	if runtime.GOOS != "windows" {
+		s, err = GenerateSpec(ctx, nil, &containers.Container{ID: t.Name()}, opts...)
+	} else {
+		s, err = GenerateSpecWithPlatform(ctx, nil, "linux/amd64", &containers.Container{ID: t.Name()}, opts...)
+	}
 	if err != nil {
 		t.Fatal(err)
 	}
